// src/models/schemas/Post.ts
// Auto-generated by Kadmium. Do not edit manually.
// Use @Kadmium.gen_skip:<name> and @Kadmium.gen_continue:<name> to protect blocks.

import { Model, ModelHookContext, ModelConfig } from "../../model/init";
import { PUBLIC_TYPE_SYMBOL, RELATIONS_SYMBOL } from "../../repo/symbols";
import { ToOneRelation, ToManyRelation } from "../../repo/types/relations";
import { KadmiumRefinementCtx } from "../../validation/types/refinement";
import { AuditFeature, RevisionsFeature } from "../../features/init";
import { User } from "./User";
import { Post_tag } from "./Post_tag";
import { Comment } from "./Comment";

export type PostSensitiveFields = never;

export type PostPublic = Post;

type __PostRelations = {
	author: ToOneRelation<User>;
	comments: ToManyRelation<Comment>;
	post_tags: ToManyRelation<Post_tag>;
};

export class Post extends Model {
	static readonly _collection = "post";
	_meta: "generated-schema" = "generated-schema";
	[PUBLIC_TYPE_SYMBOL]!: PostPublic;
	[RELATIONS_SYMBOL]!: __PostRelations;

	id!: number;
	title!: string;
	content!: string | null;
	is_published!: boolean | null;
	author_id!: string;
	published_at!: Date | null;
	created_at!: Date | null;
	updated_at!: Date | null;
// @Kadmium.gen_skip:conf
	_conf_: ModelConfig<this> = {
		hooks: {
			beforeCreate: [
				(ctx) => {
					// TODO: implement beforeCreate
				},
			],
			afterCreate: [],
			beforeUpdate: [],
			afterUpdate: [],
			beforeDelete: [],
			afterDelete: [],
			beforeRead: [],
			afterRead: [],
		},
		validation: {
			rules: [
				{ name: "example", refine: (data: this, ctx) => {
					// TODO: implement refine function
				}},
			],
		},
		features: [AuditFeature, RevisionsFeature],
	};
	// @Kadmium.gen_continue:conf
}