// src/models/schemas/Comment.ts
// Auto-generated by Kadmium. Do not edit manually.
// Use @Kadmium.gen_skip:<name> and @Kadmium.gen_continue:<name> to protect blocks.

import { Model, ModelHookContext, ModelConfig } from '../../model/index.js';
import { PUBLIC_TYPE_SYMBOL, RELATIONS_SYMBOL } from '../../repo/symbols.js';
import { ToOneRelation, ToManyRelation } from '../../repo/types/relations.js';
import { KadmiumRefinementCtx } from '../../validation/types/refinement.js';
import { SoftDeleteFeature } from '../../features/index.js';
import { Post } from './Post.js';
import { User } from './User.js';

export type CommentSensitiveFields = never;

export type CommentPublic = Comment;

type __CommentRelations = {
    post: ToOneRelation<Post>;
    user: ToOneRelation<User>;
};

export class Comment extends Model {
    static readonly _collection = 'comment';
    _meta: 'generated-schema' = 'generated-schema';
    [PUBLIC_TYPE_SYMBOL]!: CommentPublic;
    [RELATIONS_SYMBOL]!: __CommentRelations;

    id!: number;
    body!: string;
    post_id!: number;
    user_id!: string;
    deleted_at!: Date | null;
    // @Kadmium.gen_skip:conf
    _conf_: ModelConfig<this> = {
        hooks: {
            beforeCreate: [
                (ctx) => {
                    // TODO: implement beforeCreate
                },
            ],
            afterCreate: [],
            beforeUpdate: [],
            afterUpdate: [],
            beforeDelete: [],
            afterDelete: [],
            beforeRead: [],
            afterRead: [],
        },
        validation: {
            rules: [
                {
                    name: 'example',
                    refine: (data: this, ctx) => {
                        // TODO: implement refine function
                    },
                },
            ],
        },
        features: [SoftDeleteFeature],
    };
    // @Kadmium.gen_continue:conf
}