// src/models/schemas/Department.ts
// Auto-generated by Kadmium. Do not edit manually.
// Use @Kadmium.gen_skip:<name> and @Kadmium.gen_continue:<name> to protect blocks.

import { Model, ModelHookContext, ModelConfig } from "../../model/init";
import { PUBLIC_TYPE_SYMBOL, RELATIONS_SYMBOL } from "../../repo/symbols";
import { ToOneRelation, ToManyRelation } from "../../repo/types/relations";
import { KadmiumRefinementCtx } from "../../validation/types/refinement";
import { RevisionsFeature } from "../../features/init";
import { User } from "./User";

export type DepartmentSensitiveFields = never;

export type DepartmentPublic = Department;

type __DepartmentRelations = {
	users: ToManyRelation<User>;
};

export class Department extends Model {
	static readonly _collection = "department";
	_meta: "generated-schema" = "generated-schema";
	[PUBLIC_TYPE_SYMBOL]!: DepartmentPublic;
	[RELATIONS_SYMBOL]!: __DepartmentRelations;

	id!: number;
	name!: string;
// @Kadmium.gen_skip:conf
	_conf_: ModelConfig<this> = {
		hooks: {
			beforeCreate: [
				(ctx) => {
					// TODO: implement beforeCreate
				},
			],
			afterCreate: [],
			beforeUpdate: [],
			afterUpdate: [],
			beforeDelete: [],
			afterDelete: [],
			beforeRead: [],
			afterRead: [],
		},
		validation: {
			rules: [
				{ name: "example", refine: (data: this, ctx) => {
					// TODO: implement refine function
				}},
			],
		},
		features: [RevisionsFeature],
	};
	// @Kadmium.gen_continue:conf
}