// src/models/schemas/Tag.ts
// Auto-generated by Kadmium. Do not edit manually.
// Use @Kadmium.gen_skip:<name> and @Kadmium.gen_continue:<name> to protect blocks.

import { Model, ModelHookContext, ModelConfig } from "../../model/init";
import { PUBLIC_TYPE_SYMBOL, RELATIONS_SYMBOL } from "../../repo/symbols";
import { ToOneRelation, ToManyRelation } from "../../repo/types/relations";
import { KadmiumRefinementCtx } from "../../validation/types/refinement";
import { Post_tag } from "./Post_tag";

export type TagSensitiveFields = never;

export type TagPublic = Tag;

type __TagRelations = {
	post_tags: ToManyRelation<Post_tag>;
};

export class Tag extends Model {
	static readonly _collection = "tag";
	_meta: "generated-schema" = "generated-schema";
	[PUBLIC_TYPE_SYMBOL]!: TagPublic;
	[RELATIONS_SYMBOL]!: __TagRelations;

	id!: number;
	name!: string;
// @Kadmium.gen_skip:conf
	_conf_: ModelConfig<this> = {
		hooks: {
			beforeCreate: [
				(ctx) => {
					// TODO: implement beforeCreate
				},
			],
			afterCreate: [],
			beforeUpdate: [],
			afterUpdate: [],
			beforeDelete: [],
			afterDelete: [],
			beforeRead: [],
			afterRead: [],
		},
		validation: {
			rules: [
				{ name: "example", refine: (data: this, ctx) => {
					// TODO: implement refine function
				}},
			],
		},
		features: [],
	};
	// @Kadmium.gen_continue:conf
}