// src/models/schemas/Post_tag.ts
// Auto-generated by Kadmium. Do not edit manually.
// Use @Kadmium.gen_skip:<name> and @Kadmium.gen_continue:<name> to protect blocks.

import { Model, ModelHookContext, ModelConfig } from "../../model/init";
import { PUBLIC_TYPE_SYMBOL, RELATIONS_SYMBOL } from "../../repo/symbols";
import { ToOneRelation, ToManyRelation } from "../../repo/types/relations";
import { KadmiumRefinementCtx } from "../../validation/types/refinement";
import { Post } from "./Post";
import { Tag } from "./Tag";

export type Post_tagSensitiveFields = never;

export type Post_tagPublic = Post_tag;

type __Post_tagRelations = {
	post: ToOneRelation<Post>;
	tag: ToOneRelation<Tag>;
};

export class Post_tag extends Model {
	static readonly _collection = "post_tag";
	_meta: "generated-schema" = "generated-schema";
	[PUBLIC_TYPE_SYMBOL]!: Post_tagPublic;
	[RELATIONS_SYMBOL]!: __Post_tagRelations;

	id!: number;
	post_id!: number;
	tag_id!: number;
// @Kadmium.gen_skip:conf
	_conf_: ModelConfig<this> = {
		hooks: {
			beforeCreate: [
				(ctx) => {
					// TODO: implement beforeCreate
				},
			],
			afterCreate: [],
			beforeUpdate: [],
			afterUpdate: [],
			beforeDelete: [],
			afterDelete: [],
			beforeRead: [],
			afterRead: [],
		},
		validation: {
			rules: [
				{ name: "example", refine: (data: this, ctx) => {
					// TODO: implement refine function
				}},
			],
		},
		features: [],
	};
	// @Kadmium.gen_continue:conf
}